refactor(core): move deposit helpers from blocks to helpers
