+++ beacon-chain/core/helpers/deposit.go
@@ -1,4 +1,4 @@
-package blocks
+package helpers
 
 import (
 	"context"
