+++ changelog/terencechain_move-deposit-to-helpers.md
@@ -1,0 +1,2 @@
+### Ignored
+- Move deposit helpers from `beacon-chain/core/blocks` to `beacon-chain/core/helpers` (refactor only).
